support multiple networks
we now support local, ropsten and xdai, so its time to make the
html/javascript with minimal modification.
we STILL need to modify the "network" field, since in "ephemeral" mode,
the javascript code has no external way to tell which network it is
connected to.

also, style modified a bit, to look (a little) better on mobile.